+++ semcore/button/src/Button.jsx
@@ -80,7 +80,8 @@ class RootButton extends Component {
               neighborLocation={neighborLocation}
               use:theme={useTheme}
               ref={this.containerRef}
-              aria-busy={disabled || loading}
+              aria-busy={loading}
+              aria-disabled={disabled}
             >
               <SInner tag='span' loading={loading}>
                 {addonLeft ? <Button.Addon tag={addonLeft} /> : null}